+++ src/test/java/com/demo/pteam/security/login/handler/LoginAuthenticationSuccessHandlerTest.java
@@ -3,7 +3,7 @@ package com.demo.pteam.security.login.handler;
 import com.demo.pteam.authentication.domain.AccountStatus;
 import com.demo.pteam.authentication.domain.Role;
 import com.demo.pteam.global.response.ApiResponse;
-import com.demo.pteam.security.authentication.JwtService;
+import com.demo.pteam.security.jwt.JwtService;
 import com.demo.pteam.security.authentication.dto.JwtToken;
 import com.demo.pteam.security.principal.UserPrincipal;
 import com.fasterxml.jackson.databind.ObjectMapper;
